Merge

+++ Domain/DTOs/Requests/VehicleReportRequest.cs
@@ -23,7 +23,7 @@ namespace Domain.DTOs.Requests
         public DateTime ReportDate { get; set; }
         public int? GasolineLoadAmount { get; set; }
         public int? GasolineCurrentKM { get; set; }
-        public double? AmountGasoline { get; set; }
+        public decimal? AmountGasoline { get; set; }
         public int? VehicleReportUseId { get; set; }
         public List<int> Expenses { get; set; }
         public List<IFormFile> ReportImages { get; set; }
